implemented exceptions

+++ fitback/src/main/java/com/smeekens/fitback/fitback/fitback/security/services/FileStorageService.java
@@ -1,5 +1,6 @@
 package com.smeekens.fitback.fitback.fitback.security.services;
 
+import com.smeekens.fitback.fitback.fitback.exceptions.RecordNotFoundException;
 import com.smeekens.fitback.fitback.fitback.models.FileDB;
 import com.smeekens.fitback.fitback.fitback.models.User;
 import com.smeekens.fitback.fitback.fitback.repository.FileDBRepository;
@@ -11,7 +12,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Stream;
 
 
@@ -25,18 +25,22 @@ public class FileStorageService {
     private UserRepository userRepository;
 
     public FileDB uploadFile(MultipartFile multipartFile, Long id) throws IOException {
-        String fileName = StringUtils.cleanPath(Objects.requireNonNull(multipartFile.getOriginalFilename()));
-        User user = userRepository.findById(id).get();
+            String fileName = StringUtils.cleanPath(Objects.requireNonNull(multipartFile.getOriginalFilename()));
+            User user = userRepository.findById(id).get();
 
-        FileDB fileDB = new FileDB(fileName, multipartFile.getContentType(), multipartFile.getBytes());
-        fileDB.setUser(user);
-        user.addFileDB(fileDB);
-        userRepository.save(user);
-        return fileDBRepository.save(fileDB);
+            FileDB fileDB = new FileDB(fileName, multipartFile.getContentType(), multipartFile.getBytes());
+            fileDB.setUser(user);
+            user.addFileDB(fileDB);
+            userRepository.save(user);
+            return fileDBRepository.save(fileDB);
     }
 
     public FileDB getFileById(Long id) {
-        return fileDBRepository.findById(id).get();
+        if (fileDBRepository.findById(id).isPresent()) {
+            return fileDBRepository.findById(id).get();
+        } else {
+            throw new RecordNotFoundException(id);
+        }
     }
 
     public Stream<FileDB> getAllFiles() {
@@ -44,11 +48,11 @@ public class FileStorageService {
     }
 
     public void deleteFile(Long id) {
-        fileDBRepository.deleteById(id);
-    }
-
-    public Optional<FileDB> getFileDBById(Long id) {
-        return fileDBRepository.findById(id);
+        if (fileDBRepository.existsById(id)) {
+            fileDBRepository.deleteById(id);
+        } else {
+            throw new RecordNotFoundException(id);
+        }
     }
 
 }
